// Rio.Web/src/app/shared/generated/model/parcel-ownership-dto.ts
/**
 * Rio.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AccountSimpleDto } from '././account-simple-dto';
import { WaterYearDto } from '././water-year-dto';

export class ParcelOwnershipDto { 
    Account?: AccountSimpleDto;
    WaterYear?: WaterYearDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}